refactor(sbom): share the two functions both SBOM paths still duplicate
Hashing every non-test function body in the repo turns up two pieces of production logic
that exist twice, both in the two SBOM paths.

syftToDomain was the same twelve-line conversion in adapters/v1/syft_to_domain.go and
pkg/sbomscanner/v1/server.go, identical apart from one local variable name. The in-process
copy was a method on SyftAdapter that never touched its receiver, so it was a free function
in disguise. The pair had already drifted once: the sidecar's copy dropped both errors and
returned a nil the caller dereferenced (#893, fixed in #895).

packageVersion was byte-identical to internal/tools.PackageVersion, and the sidecar already
imports internal/tools for IsRateLimitError.

The conversion moves to internal/syftmeta as ToDomain. That package exists because the JSON
round-trip inside this conversion loses package metadata and Reattach is what puts it back,
so owning the round-trip and the repair together means callers get one function rather than
two steps they each have to remember to pair. Its package comment is widened to say so.

Behaviour is unchanged: the moved body is the in-process one, which the sidecar's already
matched after #895. The tests added there move with it, and adapters/v1/syft_test.go's
round-trip assertion now calls the shared function.

Re-running the scan afterwards reports zero cross-file duplicate function bodies.

Resolved #914
