+++ internal/syftmeta/reattach.go
@@ -1,11 +1,14 @@
-// Package syftmeta carries the one step that has to happen after a Syft SBOM is round-tripped
-// through JSON into the storage types: putting each package's metadata back.
+// Package syftmeta converts a Syft SBOM into the storage types, including the one step that
+// has to happen after the JSON round-trip that conversion depends on: putting each package's
+// metadata back.
 package syftmeta
 
 import (
 	"encoding/json"
 
+	"github.com/anchore/syft/syft/format/syftjson"
 	"github.com/anchore/syft/syft/format/syftjson/model"
+	"github.com/anchore/syft/syft/sbom"
 	"github.com/kubescape/go-logger"
 	"github.com/kubescape/go-logger/helpers"
 	"github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1"
@@ -49,4 +52,31 @@ func Reattach(stored []v1beta1.SyftPackage, doc []model.Package) {
 		}
 		stored[i].Metadata = b
 	}
+}
+
+// ToDomain converts a Syft SBOM into the stored document type, reattaching the metadata the
+// round-trip drops.
+//
+// Both SBOM paths need this and had a copy each: a method on SyftAdapter that never touched
+// its receiver, and a package function in the sidecar server. They had already drifted once,
+// the sidecar's dropping both errors and returning a nil the caller then dereferenced
+// (#893). Keeping one copy is the same reasoning Reattach is here for.
+func ToDomain(sbomSBOM sbom.SBOM) (*v1beta1.SyftDocument, error) {
+	doc := syftjson.ToFormatModel(sbomSBOM, syftjson.EncoderConfig{
+		Pretty: false,
+		Legacy: false,
+	})
+
+	b, err := json.Marshal(doc)
+	if err != nil {
+		return nil, err
+	}
+
+	var syftDoc *v1beta1.SyftDocument
+	if err := json.Unmarshal(b, &syftDoc); err != nil {
+		return nil, err
+	}
+	Reattach(syftDoc.Artifacts, doc.Artifacts)
+
+	return syftDoc, nil
 }
